New project with LAN Network 1
Player : Sphere
Movement : Forward only (using forces) + "glissement" (j'ai pas le mot anglais) to the right or to the left + jump
Network : LAN, with a different color for the player's view and the others.
Camera : Following the player, but still rotating with it.

Idea : stop using forces (we don't want linear movement)